add GCDWebServer reference
